US Letter paper + ToC/LoF/LoT header height adj.
Changed from a4paper to letter in the options to the report base class.
Fixed top margins while I'm at it, to keep all the head rules at the
same height.

Headers of table of contents, list of figures, and list of tables were
not previously aligned with the headers of the abstract,
acknowledgement, certification, and statement. Fix \cftbeforeZtitleskip
from -15pt to -12pt to make them align.